backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
azure-cosmos==4.5.1
pydantic==2.10.6
pydantic-settings==2.8.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.12.0
python-multipart==0.0.9
httpx>=0.24.1
pytest>=7.4.0
pytest-asyncio>=0.21.1
